% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment1_data.R
\name{get_assignment_data}
\alias{get_assignment_data}
\title{load the corresponding data set given the index number}
\usage{
get_assignment_data(index)
}
\arguments{
\item{index}{Student's index number. For eg "AS2018000"}
}
\value{
a dataframe daily summary of the Coronavirus (COVID-19) for the country associated with the index number
}
\description{
load daily summary of the Coronavirus (COVID-19) for the country associated with the index number
}
