% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment1_questions.R
\name{get_assignment_questions}
\alias{get_assignment_questions}
\title{load the corresponding data set given the index number}
\usage{
get_assignment_questions(index)
}
\arguments{
\item{index}{Student's index number with out AS}
}
\value{
a list of assignment 1 questions
}
\description{
load daily summary of the Coronavirus for the country associated with the index number
}
