% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_individual_project_country.R
\name{get_individual_project_country}
\alias{get_individual_project_country}
\title{load the corresponding data set given the index number}
\usage{
get_individual_project_country(index)
}
\arguments{
\item{index}{Student's index number. For eg "AS2018000"}
}
\value{
name of the country
}
\description{
load daily summary of the Coronavirus (COVID-19) for the country associated with the index number
}
