% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2021.R
\docType{data}
\name{index2021}
\alias{index2021}
\title{corona cases}
\format{
A data frame
\describe{
\item{V1}{index number}}
}
\source{
NA
}
\usage{
index2021
}
\description{
2021 numbers
}
\examples{
data(index2021)
head(index2021)
summary(index2021)
}
\keyword{datasets}
