% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2022.R
\docType{data}
\name{index2022}
\alias{index2022}
\title{corona cases}
\format{
A data frame
\describe{
\item{index}{index number}}
}
\source{
NA
}
\usage{
index2022
}
\description{
2022 numbers
}
\examples{
data(index2022)
head(index2022)
summary(index2022)
}
\keyword{datasets}
