% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2024.R
\docType{data}
\name{index2024}
\alias{index2024}
\title{corona cases}
\format{
A data frame
\describe{
\item{index}{index number}}
}
\source{
NA
}
\usage{
index2024
}
\description{
2024 numbers
}
\examples{
data(index2024)
head(index2024)
summary(index2024)
}
\keyword{datasets}
