% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcountry.R
\docType{data}
\name{ipcountry}
\alias{ipcountry}
\title{individual project country-wise}
\format{
A list
\describe{
\item{country}{country name}
\item{index}{index number}
}
}
\source{
NA
}
\usage{
ipcountry
}
\description{
Given the index number returns the
}
\examples{
data(ipcountry)
}
\keyword{datasets}
