% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcountry2024.R
\docType{data}
\name{ipcountry2024}
\alias{ipcountry2024}
\title{individual project country-wise}
\format{
A list
\describe{
\item{country}{country name}
\item{index}{index number}
}
}
\source{
NA
}
\usage{
ipcountry2024
}
\description{
Given the index number returns the
}
\examples{
data(ipcountry2024)
}
\keyword{datasets}
