% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sta326corona.R
\docType{data}
\name{sta326corona}
\alias{sta326corona}
\title{corona cases}
\format{
A data frame
\describe{
\item{country}{country name}
\item{date}{date}
\item{type}{An indicator for the type of cases (confirmed, death, recovered)}
\item{cases}{Number of cases on given date.}}
}
\source{
coronavirus package in R
}
\usage{
sta326corona
}
\description{
country-wise corona cases
}
\examples{
data(sta326corona)
head(sta326corona)
summary(sta326corona)
}
\keyword{datasets}
